restructured the project to include a sync namespace within the sinks namespace and introduced a directory structure
