--- C++/Logger/SyncFileSink.cpp
@@ -1,16 +1,0 @@
-#include "SyncFileSink.h"
-#include <exception>
-
-namespace Toolbox::Log::Sinks
-{
-
-SyncFileSink::SyncFileSink (const std::string & filename)
-: filename_ (filename)
-{}
-
-void SyncFileSink::output (const std::string & message)
-{
-
-}
-
-} // namespace Toolbox::Log::Sinks
