+++ C++/Logger/sinks/Sink.h
